feat: alinhamento de pedidos com o frontend e sistema de cupons
- Refatoração do `CreateOrderRequest` para aceitar a estrutura de itens do carrinho do frontend
- Implementação da lógica de aplicação de cupons de desconto no `OrdersService`
- Ajuste no mapeamento de `OrderItem` para suportar os novos DTOs de request
- Inclusão do campo `CupomId` no objeto de Pedido para rastreabilidade
- Exposição do endpoint de validação de cupons no `OrdersController`
- Padronização de tipos numéricos para `double` em itens de pedido

+++ webapi/Model/Objects/Produtos/Produto.cs
@@ -32,13 +32,13 @@ namespace DaccApi.Model
         /// Obtém ou define o preço atual do produto.
         /// </summary>
         [Column("preco")]
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
 
         /// <summary>
         /// Obtém ou define o preço original do produto (para promoções).
         /// </summary>
         [Column("preco_original")]
-        public double? PrecoOriginal { get; set; }
+        public decimal? PrecoOriginal { get; set; }
 
         /// <summary>
         /// Obtém ou define a categoria do produto.
@@ -100,7 +100,7 @@ namespace DaccApi.Model
         public bool Destaque { get; set; }
         
         [NotMapped]
-        public double AvaliacaoMedia { get; set; }
+        public decimal AvaliacaoMedia { get; set; }
         
         [NotMapped]
         public int NumeroAvaliacoes { get; set; }
@@ -226,7 +226,7 @@ namespace DaccApi.Model
         public string DiasEstimados { get; set; }
         
         [Column("custo_envio")]
-        public double? CustoEnvio { get; set; }
+        public decimal? CustoEnvio { get; set; }
         
         [Column("politica_devolucao")]
         public string PoliticaDevolucao { get; set; }
